backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydub
google-cloud-storage
google-cloud-pubsub
google-cloud-speech